backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pdfplumber==0.10.3
openpyxl==3.1.2
pandas==2.1.3
requests==2.31.0
python-dateutil==2.8.2
